+++ packages/api/src/modules/catalog/application/dtos/category/save-category.dto.ts
@@ -3,7 +3,10 @@ import * as z from 'zod'
 export namespace SaveCategoryDto {
   export const input = z.object({
     id: z.cuid2().optional(),
-    name: z.string(),
+    name: z
+      .string()
+      .min(1, 'Name is required')
+      .max(100, 'Name must be less than 100 characters'),
     description: z.string().optional(),
     image: z.url().optional(),
     parentId: z.cuid2().nullable(),