+++ packages/api/src/modules/finance/application/dtos/transaction/all-by-user.dto.ts
@@ -1,0 +1,13 @@
+import * as z from 'zod'
+
+import { TransactionEntity } from '@/modules/finance/domain/entities/transaction.entity'
+
+export namespace AllTransactionsByUserDto {
+  export const input = z.object({
+    userId: z.cuid2(),
+  })
+  export type Input = z.infer<typeof input>
+
+  export const output = z.array(z.instanceof(TransactionEntity))
+  export type Output = z.infer<typeof output>
+}