+++ apps/dashboard/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @yukinu/dashboard
 
+## 1.0.0
+
+### Major Changes
+
+- 07a4a4a: Official production-ready release featuring secure checkout, dynamic product reviews, and optimized performance.
+
+### Patch Changes
+
+- Updated dependencies [07a4a4a]
+  - @yukinu/api@1.0.0
+  - @yukinu/auth@1.0.0
+  - @yukinu/db@1.0.0
+  - @yukinu/lib@1.0.0
+  - @yukinu/ui@1.0.0
+  - @yukinu/uploadthing@1.0.0
+
+## 0.12.5
+
+### Patch Changes
+
+- Updated dependencies [9863b95]
+  - @yukinu/api@0.10.3
+
+## 0.12.4
+
+### Patch Changes
+
+- 79f806f: completed order, payment and transaction management
+
 ## 0.12.3
 
 ### Patch Changes