+++ packages/api/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @yukinu/api
 
+## 1.0.0
+
+### Major Changes
+
+- 07a4a4a: Official production-ready release featuring secure checkout, dynamic product reviews, and optimized performance.
+
+### Patch Changes
+
+- Updated dependencies [07a4a4a]
+  - @yukinu/auth@1.0.0
+  - @yukinu/db@1.0.0
+  - @yukinu/email@1.0.0
+  - @yukinu/lib@1.0.0
+  - @yukinu/uploadthing@1.0.0
+
+## 0.10.3
+
+### Patch Changes
+
+- 9863b95: added reviews product
+
 ## 0.10.2
 
 ### Patch Changes